This is the code for the credit card processing.
This code requires that you have all of your information collected from the form to be stored onto a .csv file to store all the information, and an email address for the information to be sent to you to collect. With this form you can process all the information in the fields provided. This does not allow for automated transfer of data to paypal, a bank account, etc.  